Move existing validator
We'll think about moving it into folders when we have a few more to work with.
